+++ lib/followed_list_page.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:yalkey_0206_test/yalker_profile_page.dart';
+import 'profile/yalker_profile_page.dart';
 import 'constant.dart';
 
 class FollowedListPage extends StatefulWidget {
   final int userNumber;
-  const FollowedListPage({Key? key, required this.userNumber}) : super(key: key);
+  const FollowedListPage({Key? key, required this.userNumber})
+      : super(key: key);
 
   @override
   _FollowedListPageState createState() => _FollowedListPageState();
@@ -39,7 +40,8 @@ class _FollowedListPageState extends State<FollowedListPage> {
     });
 
     FollowedListResponse followedListResponse =
-    await FollowedListResponse.fetchFollowedListResponse(widget.userNumber, _page);
+        await FollowedListResponse.fetchFollowedListResponse(
+            widget.userNumber, _page);
     if (mounted) {
       setState(() {
         _followedList
@@ -78,7 +80,6 @@ class _FollowedListPageState extends State<FollowedListPage> {
     }
   }
 
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -90,7 +91,7 @@ class _FollowedListPageState extends State<FollowedListPage> {
         onRefresh: () async {
           _clearCache();
         },
-        child:Column(
+        child: Column(
           children: <Widget>[
             Expanded(
               child: ListView.builder(
@@ -100,178 +101,240 @@ class _FollowedListPageState extends State<FollowedListPage> {
                   if (index == _followedList.length) {
                     return _loading
                         ? Container(
-                      alignment: Alignment.center,
-                      padding: EdgeInsets.all(16.0),
-                      child: CircularProgressIndicator(
-                        strokeWidth: 3.0,
-                      ),
-                    )
+                            alignment: Alignment.center,
+                            padding: EdgeInsets.all(16.0),
+                            child: CircularProgressIndicator(
+                              strokeWidth: 3.0,
+                            ),
+                          )
                         : SizedBox.shrink(); // ローディングインジケーターを表示
                   }
                   final followed = _followedList[index];
                   return InkWell(
-                    onTap: () {
-                      Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                            builder: (context) => YalkerProfilePage(userNumber: followed.followerUserNumber),
-                          )
-                      );
-                    },
-                    child:
-
-                    Padding(
-                    padding: const EdgeInsets.symmetric(
-                        vertical: 8.0, horizontal: 16.0),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: <Widget>[
-                        GestureDetector(
-                        onTap: (){
-                          print("tap");
-                        },
-                        child:Row(
+                      onTap: () {
+                        Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => YalkerProfilePage(
+                                  userNumber: followed.followerUserNumber),
+                            ));
+                      },
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(
+                            vertical: 8.0, horizontal: 16.0),
+                        child: Column(
                           crossAxisAlignment: CrossAxisAlignment.start,
                           children: <Widget>[
-                            if (followed.followerIconimage=="") const CircleAvatar(
-                              backgroundColor: Colors.white,
-                              backgroundImage: NetworkImage(
-                                'https://yalkey-s3.s3.ap-southeast-2.amazonaws.com/static/img/user.png',
-                              ),
-                            ),
-                            if (followed.followerIconimage!="") CircleAvatar(
-                              backgroundImage: NetworkImage(
-                                'https://yalkey-s3.s3.ap-southeast-2.amazonaws.com/media/iconimage/${followed.followerIconimage}',
-                              ),
-                            ),
-                            SizedBox(width: 16.0),
-                            Expanded(
-                              child: Column(
+                            GestureDetector(
+                              onTap: () {
+                                print("tap");
+                              },
+                              child: Row(
                                 crossAxisAlignment: CrossAxisAlignment.start,
                                 children: <Widget>[
-                                  Text(
-                                    followed.followerName,
-                                    style: const TextStyle(fontSize: 20.0),
+                                  if (followed.followerIconimage == "")
+                                    const CircleAvatar(
+                                      backgroundColor: Colors.white,
+                                      backgroundImage: NetworkImage(
+                                        'https://yalkey-s3.s3.ap-southeast-2.amazonaws.com/static/img/user.png',
+                                      ),
+                                    ),
+                                  if (followed.followerIconimage != "")
+                                    CircleAvatar(
+                                      backgroundImage: NetworkImage(
+                                        'https://yalkey-s3.s3.ap-southeast-2.amazonaws.com/media/iconimage/${followed.followerIconimage}',
+                                      ),
+                                    ),
+                                  SizedBox(width: 16.0),
+                                  Expanded(
+                                    child: Column(
+                                      crossAxisAlignment:
+                                          CrossAxisAlignment.start,
+                                      children: <Widget>[
+                                        Text(
+                                          followed.followerName,
+                                          style:
+                                              const TextStyle(fontSize: 20.0),
+                                        ),
+                                        Row(
+                                          children: [
+                                            if (followed.followerPrivate)
+                                              const Icon(
+                                                Icons.lock,
+                                                color: Colors.grey,
+                                                size: 12.0,
+                                              ),
+                                            Text(
+                                              '@${followed.followerUserId}',
+                                              style: const TextStyle(
+                                                  fontSize: 12.0,
+                                                  color: Colors.grey),
+                                            ),
+                                          ],
+                                        ),
+                                        SizedBox(height: 4.0),
+                                        Row(
+                                          children: [
+                                            if (followed.followerSuperEarlyBird)
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                        horizontal: 3,
+                                                        vertical: 1),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    color:
+                                                        const Color(0xFFAE0103),
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            8),
+                                                  ),
+                                                  child: const Padding(
+                                                    padding:
+                                                        EdgeInsets.symmetric(
+                                                            horizontal: 3,
+                                                            vertical: 1),
+                                                    child: Text(
+                                                      "超早起き",
+                                                      style: TextStyle(
+                                                          fontSize: 10.0,
+                                                          color: Colors.white),
+                                                    ),
+                                                  ),
+                                                ),
+                                              ),
+                                            if (followed.followerSuperEarlyBird)
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                        horizontal: 3,
+                                                        vertical: 1),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    color:
+                                                        const Color(0xFFAE0103),
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            8),
+                                                  ),
+                                                  child: const Padding(
+                                                    padding:
+                                                        EdgeInsets.symmetric(
+                                                            horizontal: 3,
+                                                            vertical: 1),
+                                                    child: Text(
+                                                      "早起き",
+                                                      style: TextStyle(
+                                                          fontSize: 10.0,
+                                                          color: Colors.white),
+                                                    ),
+                                                  ),
+                                                ),
+                                              ),
+                                            if (followed
+                                                .followerSuperHardWorker)
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                        horizontal: 3,
+                                                        vertical: 1),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    color:
+                                                        const Color(0xFFAE0103),
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            8),
+                                                  ),
+                                                  child: const Padding(
+                                                    padding:
+                                                        EdgeInsets.symmetric(
+                                                            horizontal: 3,
+                                                            vertical: 1),
+                                                    child: Text(
+                                                      "超努力家",
+                                                      style: TextStyle(
+                                                          fontSize: 10.0,
+                                                          color: Colors.white),
+                                                    ),
+                                                  ),
+                                                ),
+                                              ),
+                                            if (followed.followerHardWorker)
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                        horizontal: 3,
+                                                        vertical: 1),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    color:
+                                                        const Color(0xFFAE0103),
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            8),
+                                                  ),
+                                                  child: const Padding(
+                                                    padding:
+                                                        EdgeInsets.symmetric(
+                                                            horizontal: 3,
+                                                            vertical: 1),
+                                                    child: Text(
+                                                      "努力家",
+                                                      style: TextStyle(
+                                                          fontSize: 10.0,
+                                                          color: Colors.white),
+                                                    ),
+                                                  ),
+                                                ),
+                                              ),
+                                            if (followed
+                                                .followerRegularCustomer)
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                        horizontal: 3,
+                                                        vertical: 1),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    color:
+                                                        const Color(0xFFAE0103),
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            8),
+                                                  ),
+                                                  child: const Padding(
+                                                    padding:
+                                                        EdgeInsets.symmetric(
+                                                            horizontal: 3,
+                                                            vertical: 1),
+                                                    child: Text(
+                                                      "常連",
+                                                      style: TextStyle(
+                                                          fontSize: 10.0,
+                                                          color: Colors.white),
+                                                    ),
+                                                  ),
+                                                ),
+                                              ),
+                                          ],
+                                        ),
+                                        followed.followerProfile != ''
+                                            ? Text(
+                                                followed.followerProfile,
+                                                style:
+                                                    TextStyle(fontSize: 14.0),
+                                              )
+                                            : SizedBox.shrink(),
+                                      ],
+                                    ),
                                   ),
-                                  Row(
-                                    children: [
-                                      if (followed.followerPrivate) const Icon(
-                                        Icons.lock,
-                                        color: Colors.grey,
-                                        size: 12.0,
-                                      ),
-                                      Text(
-                                        '@${followed.followerUserId}',
-                                        style: const TextStyle(fontSize: 12.0, color: Colors.grey),
-                                      ),
-                                    ],
-                                  ),
-                                  SizedBox(height: 4.0),
-                                  Row(
-                                    children: [
-                                      if (followed.followerSuperEarlyBird) Padding(
-                                          padding: const EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                          child: Container(
-                                          decoration: BoxDecoration(
-                                            color: const Color(0xFFAE0103),
-                                            borderRadius: BorderRadius.circular(8),
-                                          ),
-                                          child: const Padding(
-                                            padding: EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                            child: Text(
-                                              "超早起き",
-                                              style: TextStyle(fontSize: 10.0, color: Colors.white),
-                                            ),
-                                          ),
-                                        ),
-                                      ),
-                                      if (followed.followerSuperEarlyBird) Padding(
-                                        padding: const EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                        child: Container(
-                                          decoration: BoxDecoration(
-                                            color: const Color(0xFFAE0103),
-                                            borderRadius: BorderRadius.circular(8),
-                                          ),
-                                          child: const Padding(
-                                            padding: EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                            child: Text(
-                                              "早起き",
-                                              style: TextStyle(fontSize: 10.0, color: Colors.white),
-                                            ),
-                                          ),
-                                        ),
-                                      ),
-                                      if (followed.followerSuperHardWorker) Padding(
-                                        padding: const EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                        child: Container(
-                                          decoration: BoxDecoration(
-                                            color: const Color(0xFFAE0103),
-                                            borderRadius: BorderRadius.circular(8),
-                                          ),
-                                          child: const Padding(
-                                            padding: EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                            child: Text(
-                                              "超努力家",
-                                              style: TextStyle(fontSize: 10.0, color: Colors.white),
-                                            ),
-                                          ),
-                                        ),
-                                      ),
-                                      if (followed.followerHardWorker) Padding(
-                                        padding: const EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                        child: Container(
-                                          decoration: BoxDecoration(
-                                            color: const Color(0xFFAE0103),
-                                            borderRadius: BorderRadius.circular(8),
-                                          ),
-                                          child: const Padding(
-                                            padding: EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                            child: Text(
-                                              "努力家",
-                                              style: TextStyle(fontSize: 10.0, color: Colors.white),
-                                            ),
-                                          ),
-                                        ),
-                                      ),
-                                      if (followed.followerRegularCustomer) Padding(
-                                        padding: const EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                        child: Container(
-                                          decoration: BoxDecoration(
-                                            color: const Color(0xFFAE0103),
-                                            borderRadius: BorderRadius.circular(8),
-                                          ),
-                                          child: const Padding(
-                                            padding: EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                            child: Text(
-                                              "常連",
-                                              style: TextStyle(fontSize: 10.0, color: Colors.white),
-                                            ),
-                                          ),
-                                        ),
-                                      ),
-                                    ],
-                                  ),
-
-
-
-                                  followed.followerProfile != ''
-                                      ? Text(
-                                    followed.followerProfile,
-                                    style: TextStyle(fontSize: 14.0),
-                                  )
-                                      : SizedBox.shrink(),
                                 ],
                               ),
                             ),
                           ],
                         ),
-                        ),
-                      ],
-                    ),
-                  ));
-
-
-
+                      ));
                 },
               ),
             ),
