+++ lib/following_list_page.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:yalkey_0206_test/yalker_profile_page.dart';
+import 'profile/yalker_profile_page.dart';
 import 'constant.dart';
 
 class FollowingListPage extends StatefulWidget {
   final int userNumber;
-  const FollowingListPage({Key? key, required this.userNumber}) : super(key: key);
+  const FollowingListPage({Key? key, required this.userNumber})
+      : super(key: key);
 
   @override
   _FollowingListPageState createState() => _FollowingListPageState();
@@ -39,7 +40,8 @@ class _FollowingListPageState extends State<FollowingListPage> {
     });
 
     FollowingListResponse followingListResponse =
-    await FollowingListResponse.fetchFollowingListResponse(widget.userNumber, _page);
+        await FollowingListResponse.fetchFollowingListResponse(
+            widget.userNumber, _page);
     if (mounted) {
       setState(() {
         _followingList
@@ -78,7 +80,6 @@ class _FollowingListPageState extends State<FollowingListPage> {
     }
   }
 
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -90,181 +91,253 @@ class _FollowingListPageState extends State<FollowingListPage> {
         onRefresh: () async {
           _clearCache();
         },
-        child:Column(
+        child: Column(
           children: <Widget>[
             Expanded(
               child: ListView.builder(
                 controller: _scrollController, // スクロールコントローラーを設定
-                itemCount: _followingList.length + 1, // リストアイテム数 + ローディングインジケーター
+                itemCount:
+                    _followingList.length + 1, // リストアイテム数 + ローディングインジケーター
                 itemBuilder: (context, index) {
                   if (index == _followingList.length) {
                     return _loading
                         ? Container(
-                      alignment: Alignment.center,
-                      padding: EdgeInsets.all(16.0),
-                      child: CircularProgressIndicator(
-                        strokeWidth: 3.0,
-                      ),
-                    )
+                            alignment: Alignment.center,
+                            padding: EdgeInsets.all(16.0),
+                            child: CircularProgressIndicator(
+                              strokeWidth: 3.0,
+                            ),
+                          )
                         : SizedBox.shrink(); // ローディングインジケーターを表示
                   }
                   final following = _followingList[index];
                   return InkWell(
                       onTap: () {
-                    Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => YalkerProfilePage(userNumber: following.followedUserNumber),
-                        )
-                    );
-                  },
-                  child: Padding(
-                    padding: const EdgeInsets.symmetric(
-                        vertical: 8.0, horizontal: 16.0),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: <Widget>[
-                        GestureDetector(
-                          onTap: (){
-                            print("tap");
-                          },
-                          child:Row(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              if (following.followedIconimage=="") const CircleAvatar(
-                                backgroundColor: Colors.white,
-                                backgroundImage: NetworkImage(
-                                  'https://yalkey-s3.s3.ap-southeast-2.amazonaws.com/static/img/user.png',
-                                ),
-                              ),
-                              if (following.followedIconimage!="") CircleAvatar(
-                                backgroundImage: NetworkImage(
-                                  'https://yalkey-s3.s3.ap-southeast-2.amazonaws.com/media/iconimage/${following.followedIconimage}',
-                                ),
-                              ),
-                              SizedBox(width: 16.0),
-                              Expanded(
-                                child: Column(
-                                  crossAxisAlignment: CrossAxisAlignment.start,
-                                  children: <Widget>[
-                                    Text(
-                                      following.followedName,
-                                      style: const TextStyle(fontSize: 20.0),
+                        Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => YalkerProfilePage(
+                                  userNumber: following.followedUserNumber),
+                            ));
+                      },
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(
+                            vertical: 8.0, horizontal: 16.0),
+                        child: Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: <Widget>[
+                            GestureDetector(
+                              onTap: () {
+                                print("tap");
+                              },
+                              child: Row(
+                                crossAxisAlignment: CrossAxisAlignment.start,
+                                children: <Widget>[
+                                  if (following.followedIconimage == "")
+                                    const CircleAvatar(
+                                      backgroundColor: Colors.white,
+                                      backgroundImage: NetworkImage(
+                                        'https://yalkey-s3.s3.ap-southeast-2.amazonaws.com/static/img/user.png',
+                                      ),
                                     ),
-
-                                    Row(
-                                      children: [
-                                        if (following.followedPrivate) const Icon(
-                                          Icons.lock,
-                                          color: Colors.grey,
-                                          size: 12.0,
-                                        ),
+                                  if (following.followedIconimage != "")
+                                    CircleAvatar(
+                                      backgroundImage: NetworkImage(
+                                        'https://yalkey-s3.s3.ap-southeast-2.amazonaws.com/media/iconimage/${following.followedIconimage}',
+                                      ),
+                                    ),
+                                  SizedBox(width: 16.0),
+                                  Expanded(
+                                    child: Column(
+                                      crossAxisAlignment:
+                                          CrossAxisAlignment.start,
+                                      children: <Widget>[
                                         Text(
-                                          '@${following.followedUserId}',
-                                          style: const TextStyle(fontSize: 12.0, color: Colors.grey),
+                                          following.followedName,
+                                          style:
+                                              const TextStyle(fontSize: 20.0),
                                         ),
+                                        Row(
+                                          children: [
+                                            if (following.followedPrivate)
+                                              const Icon(
+                                                Icons.lock,
+                                                color: Colors.grey,
+                                                size: 12.0,
+                                              ),
+                                            Text(
+                                              '@${following.followedUserId}',
+                                              style: const TextStyle(
+                                                  fontSize: 12.0,
+                                                  color: Colors.grey),
+                                            ),
+                                          ],
+                                        ),
+                                        SizedBox(height: 4.0),
+                                        Row(
+                                          children: [
+                                            if (following
+                                                .followedSuperEarlyBird)
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                        horizontal: 3,
+                                                        vertical: 1),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    color:
+                                                        const Color(0xFFAE0103),
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            8),
+                                                  ),
+                                                  child: const Padding(
+                                                    padding:
+                                                        EdgeInsets.symmetric(
+                                                            horizontal: 3,
+                                                            vertical: 1),
+                                                    child: Text(
+                                                      "超早起き",
+                                                      style: TextStyle(
+                                                          fontSize: 10.0,
+                                                          color: Colors.white),
+                                                    ),
+                                                  ),
+                                                ),
+                                              ),
+                                            if (following
+                                                .followedSuperEarlyBird)
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                        horizontal: 3,
+                                                        vertical: 1),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    color:
+                                                        const Color(0xFFAE0103),
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            8),
+                                                  ),
+                                                  child: const Padding(
+                                                    padding:
+                                                        EdgeInsets.symmetric(
+                                                            horizontal: 3,
+                                                            vertical: 1),
+                                                    child: Text(
+                                                      "早起き",
+                                                      style: TextStyle(
+                                                          fontSize: 10.0,
+                                                          color: Colors.white),
+                                                    ),
+                                                  ),
+                                                ),
+                                              ),
+                                            if (following
+                                                .followedSuperHardWorker)
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                        horizontal: 3,
+                                                        vertical: 1),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    color:
+                                                        const Color(0xFFAE0103),
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            8),
+                                                  ),
+                                                  child: const Padding(
+                                                    padding:
+                                                        EdgeInsets.symmetric(
+                                                            horizontal: 3,
+                                                            vertical: 1),
+                                                    child: Text(
+                                                      "超努力家",
+                                                      style: TextStyle(
+                                                          fontSize: 10.0,
+                                                          color: Colors.white),
+                                                    ),
+                                                  ),
+                                                ),
+                                              ),
+                                            if (following.followedHardWorker)
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                        horizontal: 3,
+                                                        vertical: 1),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    color:
+                                                        const Color(0xFFAE0103),
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            8),
+                                                  ),
+                                                  child: const Padding(
+                                                    padding:
+                                                        EdgeInsets.symmetric(
+                                                            horizontal: 3,
+                                                            vertical: 1),
+                                                    child: Text(
+                                                      "努力家",
+                                                      style: TextStyle(
+                                                          fontSize: 10.0,
+                                                          color: Colors.white),
+                                                    ),
+                                                  ),
+                                                ),
+                                              ),
+                                            if (following
+                                                .followedRegularCustomer)
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                        horizontal: 3,
+                                                        vertical: 1),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    color:
+                                                        const Color(0xFFAE0103),
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            8),
+                                                  ),
+                                                  child: const Padding(
+                                                    padding:
+                                                        EdgeInsets.symmetric(
+                                                            horizontal: 3,
+                                                            vertical: 1),
+                                                    child: Text(
+                                                      "常連",
+                                                      style: TextStyle(
+                                                          fontSize: 10.0,
+                                                          color: Colors.white),
+                                                    ),
+                                                  ),
+                                                ),
+                                              ),
+                                          ],
+                                        ),
+                                        following.followedProfile != ''
+                                            ? Text(
+                                                following.followedProfile,
+                                                style:
+                                                    TextStyle(fontSize: 14.0),
+                                              )
+                                            : SizedBox.shrink(),
                                       ],
                                     ),
-                                    SizedBox(height: 4.0),
-                                    Row(
-                                      children: [
-                                        if (following.followedSuperEarlyBird) Padding(
-                                          padding: const EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                          child: Container(
-                                            decoration: BoxDecoration(
-                                              color: const Color(0xFFAE0103),
-                                              borderRadius: BorderRadius.circular(8),
-                                            ),
-                                            child: const Padding(
-                                              padding: EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                              child: Text(
-                                                "超早起き",
-                                                style: TextStyle(fontSize: 10.0, color: Colors.white),
-                                              ),
-                                            ),
-                                          ),
-                                        ),
-                                        if (following.followedSuperEarlyBird) Padding(
-                                          padding: const EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                          child: Container(
-                                            decoration: BoxDecoration(
-                                              color: const Color(0xFFAE0103),
-                                              borderRadius: BorderRadius.circular(8),
-                                            ),
-                                            child: const Padding(
-                                              padding: EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                              child: Text(
-                                                "早起き",
-                                                style: TextStyle(fontSize: 10.0, color: Colors.white),
-                                              ),
-                                            ),
-                                          ),
-                                        ),
-                                        if (following.followedSuperHardWorker) Padding(
-                                          padding: const EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                          child: Container(
-                                            decoration: BoxDecoration(
-                                              color: const Color(0xFFAE0103),
-                                              borderRadius: BorderRadius.circular(8),
-                                            ),
-                                            child: const Padding(
-                                              padding: EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                              child: Text(
-                                                "超努力家",
-                                                style: TextStyle(fontSize: 10.0, color: Colors.white),
-                                              ),
-                                            ),
-                                          ),
-                                        ),
-                                        if (following.followedHardWorker) Padding(
-                                          padding: const EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                          child: Container(
-                                            decoration: BoxDecoration(
-                                              color: const Color(0xFFAE0103),
-                                              borderRadius: BorderRadius.circular(8),
-                                            ),
-                                            child: const Padding(
-                                              padding: EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                              child: Text(
-                                                "努力家",
-                                                style: TextStyle(fontSize: 10.0, color: Colors.white),
-                                              ),
-                                            ),
-                                          ),
-                                        ),
-                                        if (following.followedRegularCustomer) Padding(
-                                          padding: const EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                          child: Container(
-                                            decoration: BoxDecoration(
-                                              color: const Color(0xFFAE0103),
-                                              borderRadius: BorderRadius.circular(8),
-                                            ),
-                                            child: const Padding(
-                                              padding: EdgeInsets.symmetric(horizontal:3, vertical: 1),
-                                              child: Text(
-                                                "常連",
-                                                style: TextStyle(fontSize: 10.0, color: Colors.white),
-                                              ),
-                                            ),
-                                          ),
-                                        ),
-                                      ],
-                                    ),
-                                    following.followedProfile != ''
-                                        ? Text(
-                                      following.followedProfile,
-                                      style: TextStyle(fontSize: 14.0),
-                                    )
-                                        : SizedBox.shrink(),
-                                  ],
-                                ),
+                                  ),
+                                ],
                               ),
-                            ],
-                          ),
+                            ),
+                          ],
                         ),
-                      ],
-                    ),
-                  ));
+                      ));
                 },
               ),
             ),
