+++ lib/profile/yalker_profile_page.dart
@@ -1,12 +1,12 @@
 import 'dart:convert';
 import 'package:flutter/material.dart';
-import 'api.dart';
+import '../api.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'followed_list_page.dart';
-import 'following_list_page.dart';
+import '../followed_list_page.dart';
+import '../following_list_page.dart';
 import 'profile_edit_page.dart';
 import 'bar_graph.dart';
-import 'constant.dart';
+import '../constant.dart';
 import 'achievement_calendar.dart';
 
 const Map<String, String> badge2Explanation = {
@@ -34,21 +34,19 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
   late bool _loadingReportList = false;
   int relationType = 1;
 
-
   @override
   void initState() {
     super.initState();
     _fetchProfileData();
     _fetchReportList();
-    if (_profileData != null){
-      if(_profileData!['yalker_profile']['relation_type']!=null){
+    if (_profileData != null) {
+      if (_profileData!['yalker_profile']['relation_type'] != null) {
         relationType = _profileData!['yalker_profile']['relation_type'];
       }
-    }else{
+    } else {
       relationType = 1;
     }
   }
-
 
   Future<void> _fetchProfileData() async {
     try {
@@ -111,9 +109,8 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
     );
   }
 
-  void _onPressedButton(){
-    setState((){
-
+  void _onPressedButton() {
+    setState(() {
       switch (relationType) {
         case 1:
           relationType++;
@@ -130,8 +127,6 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
       }
     });
   }
-
-
 
   @override
   Widget build(BuildContext context) {
@@ -153,19 +148,19 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                         crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
                           _profileData!['yalker_profile']['iconimage'] == ""
-                            ? const CircleAvatar(
-                              backgroundColor: Colors.white,
-                              backgroundImage: NetworkImage(
-                                'https://yalkey-s3.s3.ap-southeast-2.amazonaws.com/static/img/user.png',
-                              ),
-                              radius: 25,
-                            )
-                                : CircleAvatar(
-                              backgroundImage: NetworkImage(
-                                '${_profileData!['yalker_profile']['iconimage']}',
-                              ),
-                              radius: 25,
-                            ),
+                              ? const CircleAvatar(
+                                  backgroundColor: Colors.white,
+                                  backgroundImage: NetworkImage(
+                                    'https://yalkey-s3.s3.ap-southeast-2.amazonaws.com/static/img/user.png',
+                                  ),
+                                  radius: 25,
+                                )
+                              : CircleAvatar(
+                                  backgroundImage: NetworkImage(
+                                    '${_profileData!['yalker_profile']['iconimage']}',
+                                  ),
+                                  radius: 25,
+                                ),
                           SizedBox(width: 10), // アイコンと名前の間隔を設定
                           Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
@@ -176,7 +171,8 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                               ),
                               Row(
                                 children: [
-                                  if (_profileData!['yalker_profile']['lock'] ?? false)
+                                  if (_profileData!['yalker_profile']['lock'] ??
+                                      false)
                                     const Icon(
                                       Icons.lock,
                                       color: Colors.grey,
@@ -185,41 +181,34 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                                   Text(
                                     '@${_profileData!['yalker_profile']['user_id']}',
                                     style: const TextStyle(
-                                        fontSize: 16.0,
-                                        color: Colors.grey),
+                                        fontSize: 16.0, color: Colors.grey),
                                   ),
                                 ],
                               ),
                               const SizedBox(height: 4.0),
                               Row(
                                 children: [
-                                  if (_profileData!['yalker_profile']['super_early_bird'] ??
+                                  if (_profileData!['yalker_profile']
+                                          ['super_early_bird'] ??
                                       false)
                                     GestureDetector(
-                                      behavior:
-                                      HitTestBehavior.translucent,
+                                      behavior: HitTestBehavior.translucent,
                                       onTap: () {
                                         _showExplanation(context,
                                             badge2Explanation['超早起き']!);
                                       },
                                       child: Padding(
-                                        padding:
-                                        const EdgeInsets.symmetric(
-                                            horizontal: 3,
-                                            vertical: 1),
+                                        padding: const EdgeInsets.symmetric(
+                                            horizontal: 3, vertical: 1),
                                         child: Container(
                                           decoration: BoxDecoration(
-                                            color:
-                                            const Color(0xFFAE0103),
+                                            color: const Color(0xFFAE0103),
                                             borderRadius:
-                                            BorderRadius.circular(
-                                                8),
+                                                BorderRadius.circular(8),
                                           ),
                                           child: const Padding(
-                                            padding:
-                                            EdgeInsets.symmetric(
-                                                horizontal: 3,
-                                                vertical: 1),
+                                            padding: EdgeInsets.symmetric(
+                                                horizontal: 3, vertical: 1),
                                             child: Text(
                                               "超早起き",
                                               style: TextStyle(
@@ -233,30 +222,23 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                                   if (_profileData!['yalker_profile']['lock'] ??
                                       false)
                                     GestureDetector(
-                                      behavior:
-                                      HitTestBehavior.translucent,
+                                      behavior: HitTestBehavior.translucent,
                                       onTap: () {
                                         _showExplanation(context,
                                             badge2Explanation['early_bird']!);
                                       },
                                       child: Padding(
-                                        padding:
-                                        const EdgeInsets.symmetric(
-                                            horizontal: 3,
-                                            vertical: 1),
+                                        padding: const EdgeInsets.symmetric(
+                                            horizontal: 3, vertical: 1),
                                         child: Container(
                                           decoration: BoxDecoration(
-                                            color:
-                                            const Color(0xFFAE0103),
+                                            color: const Color(0xFFAE0103),
                                             borderRadius:
-                                            BorderRadius.circular(
-                                                8),
+                                                BorderRadius.circular(8),
                                           ),
                                           child: const Padding(
-                                            padding:
-                                            EdgeInsets.symmetric(
-                                                horizontal: 3,
-                                                vertical: 1),
+                                            padding: EdgeInsets.symmetric(
+                                                horizontal: 3, vertical: 1),
                                             child: Text(
                                               "早起き",
                                               style: TextStyle(
@@ -267,33 +249,27 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                                         ),
                                       ),
                                     ),
-                                  if (_profileData!['yalker_profile']['super_hard_worker'] ??
+                                  if (_profileData!['yalker_profile']
+                                          ['super_hard_worker'] ??
                                       false)
                                     GestureDetector(
-                                      behavior:
-                                      HitTestBehavior.translucent,
+                                      behavior: HitTestBehavior.translucent,
                                       onTap: () {
                                         _showExplanation(context,
                                             badge2Explanation['超努力家']!);
                                       },
                                       child: Padding(
-                                        padding:
-                                        const EdgeInsets.symmetric(
-                                            horizontal: 3,
-                                            vertical: 1),
+                                        padding: const EdgeInsets.symmetric(
+                                            horizontal: 3, vertical: 1),
                                         child: Container(
                                           decoration: BoxDecoration(
-                                            color:
-                                            const Color(0xFFAE0103),
+                                            color: const Color(0xFFAE0103),
                                             borderRadius:
-                                            BorderRadius.circular(
-                                                8),
+                                                BorderRadius.circular(8),
                                           ),
                                           child: const Padding(
-                                            padding:
-                                            EdgeInsets.symmetric(
-                                                horizontal: 3,
-                                                vertical: 1),
+                                            padding: EdgeInsets.symmetric(
+                                                horizontal: 3, vertical: 1),
                                             child: Text(
                                               "超努力家",
                                               style: TextStyle(
@@ -304,33 +280,27 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                                         ),
                                       ),
                                     ),
-                                  if (_profileData!['yalker_profile']['hard_worker'] ??
+                                  if (_profileData!['yalker_profile']
+                                          ['hard_worker'] ??
                                       false)
                                     GestureDetector(
-                                      behavior:
-                                      HitTestBehavior.translucent,
+                                      behavior: HitTestBehavior.translucent,
                                       onTap: () {
-                                        _showExplanation(context,
-                                            badge2Explanation['努力家']!);
+                                        _showExplanation(
+                                            context, badge2Explanation['努力家']!);
                                       },
                                       child: Padding(
-                                        padding:
-                                        const EdgeInsets.symmetric(
-                                            horizontal: 3,
-                                            vertical: 1),
+                                        padding: const EdgeInsets.symmetric(
+                                            horizontal: 3, vertical: 1),
                                         child: Container(
                                           decoration: BoxDecoration(
-                                            color:
-                                            const Color(0xFFAE0103),
+                                            color: const Color(0xFFAE0103),
                                             borderRadius:
-                                            BorderRadius.circular(
-                                                8),
+                                                BorderRadius.circular(8),
                                           ),
                                           child: const Padding(
-                                            padding:
-                                            EdgeInsets.symmetric(
-                                                horizontal: 3,
-                                                vertical: 1),
+                                            padding: EdgeInsets.symmetric(
+                                                horizontal: 3, vertical: 1),
                                             child: Text(
                                               "努力家",
                                               style: TextStyle(
@@ -341,33 +311,27 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                                         ),
                                       ),
                                     ),
-                                  if (_profileData!['yalker_profile']['regular_customer'] ??
+                                  if (_profileData!['yalker_profile']
+                                          ['regular_customer'] ??
                                       false)
                                     GestureDetector(
-                                      behavior:
-                                      HitTestBehavior.translucent,
+                                      behavior: HitTestBehavior.translucent,
                                       onTap: () {
-                                        _showExplanation(context,
-                                            badge2Explanation['常連']!);
+                                        _showExplanation(
+                                            context, badge2Explanation['常連']!);
                                       },
                                       child: Padding(
-                                        padding:
-                                        const EdgeInsets.symmetric(
-                                            horizontal: 3,
-                                            vertical: 1),
+                                        padding: const EdgeInsets.symmetric(
+                                            horizontal: 3, vertical: 1),
                                         child: Container(
                                           decoration: BoxDecoration(
-                                            color:
-                                            const Color(0xFFAE0103),
+                                            color: const Color(0xFFAE0103),
                                             borderRadius:
-                                            BorderRadius.circular(
-                                                8),
+                                                BorderRadius.circular(8),
                                           ),
                                           child: const Padding(
-                                            padding:
-                                            EdgeInsets.symmetric(
-                                                horizontal: 3,
-                                                vertical: 1),
+                                            padding: EdgeInsets.symmetric(
+                                                horizontal: 3, vertical: 1),
                                             child: Text(
                                               "常連",
                                               style: TextStyle(
@@ -387,55 +351,62 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                         ],
                       ),
                       SizedBox(height: 5),
-                      Row(
-                          children: [
-                            GestureDetector( //InkWellでも同じ
-                              onTap: () {
-                                Navigator.push(
-                                  context,
-                                  MaterialPageRoute(
-                                    builder: (context) => FollowingListPage(userNumber: _profileData!['yalker_profile']['user_number']),
-                                  ),
-                                );
-                              },
-                              child: Text(
-                                'フォロー ',
-                                textAlign: TextAlign.start, // 左詰めに設定
+                      Row(children: [
+                        GestureDetector(
+                          //InkWellでも同じ
+                          onTap: () {
+                            Navigator.push(
+                              context,
+                              MaterialPageRoute(
+                                builder: (context) => FollowingListPage(
+                                    userNumber: _profileData!['yalker_profile']
+                                        ['user_number']),
                               ),
-                            ),
-                            SizedBox(width: 20),
-                            GestureDetector( //InkWellでも同じ
-                              onTap: () {
-                                Navigator.push(
-                                  context,
-                                  MaterialPageRoute(
-                                    builder: (context) => FollowedListPage(userNumber: _profileData!['yalker_profile']['user_number']),
-                                  ),
-                                );
-                              },
-                              child: Text(
-                                'フォロワー ',
-                                textAlign: TextAlign.start, // 左詰めに設定
+                            );
+                          },
+                          child: Text(
+                            'フォロー ',
+                            textAlign: TextAlign.start, // 左詰めに設定
+                          ),
+                        ),
+                        SizedBox(width: 20),
+                        GestureDetector(
+                          //InkWellでも同じ
+                          onTap: () {
+                            Navigator.push(
+                              context,
+                              MaterialPageRoute(
+                                builder: (context) => FollowedListPage(
+                                    userNumber: _profileData!['yalker_profile']
+                                        ['user_number']),
                               ),
-                            ),
-                          ]
-                      ),
+                            );
+                          },
+                          child: Text(
+                            'フォロワー ',
+                            textAlign: TextAlign.start, // 左詰めに設定
+                          ),
+                        ),
+                      ]),
                       SizedBox(height: 20), // 余白を追加
                       Text(
                         '${_profileData!['yalker_profile']['profile']}',
                         textAlign: TextAlign.start, // 左詰めに設定
                       ),
                       SizedBox(height: 10),
-                      if (relationType==0)
+                      if (relationType == 0)
                         ElevatedButton(
                           onPressed: _editProfile, // 編集ボタンが押された時の処理
                           child: Text('編集'),
                         ),
-                      if (relationType==1)
+                      if (relationType == 1)
                         ElevatedButton(
-                          onPressed: (){
+                          onPressed: () {
                             try {
-                              httpPost('follow/${_profileData!['yalker_profile']['user_number']}', {"email":"email"},jwt: true);
+                              httpPost(
+                                  'follow/${_profileData!['yalker_profile']['user_number']}',
+                                  {"email": "email"},
+                                  jwt: true);
                               _onPressedButton();
                               print(relationType);
                             } catch (error) {
@@ -444,11 +415,14 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                           },
                           child: Text('フォローする'),
                         ),
-                      if (relationType==2)
+                      if (relationType == 2)
                         ElevatedButton(
-                          onPressed: (){
+                          onPressed: () {
                             try {
-                              httpPost('unfollow/${_profileData!['yalker_profile']['user_number']}', {"email":"email"},jwt: true);
+                              httpPost(
+                                  'unfollow/${_profileData!['yalker_profile']['user_number']}',
+                                  {"email": "email"},
+                                  jwt: true);
                               _onPressedButton();
                             } catch (error) {
                               print('Error unfollow: $error');
@@ -456,11 +430,14 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                           },
                           child: Text('フォロー解除'),
                         ),
-                      if (relationType==3)
+                      if (relationType == 3)
                         ElevatedButton(
-                          onPressed: (){
+                          onPressed: () {
                             try {
-                              httpPost('followrequest/${_profileData!['yalker_profile']['user_number']}', {"email":"email"},jwt: true);
+                              httpPost(
+                                  'followrequest/${_profileData!['yalker_profile']['user_number']}',
+                                  {"email": "email"},
+                                  jwt: true);
                               _onPressedButton();
                             } catch (error) {
                               print('Error follow request: $error');
@@ -468,11 +445,14 @@ class _YalkerProfilePageState extends State<YalkerProfilePage> {
                           },
                           child: Text('リクエスト'),
                         ),
-                      if (relationType==4)
+                      if (relationType == 4)
                         ElevatedButton(
-                          onPressed: (){
+                          onPressed: () {
                             try {
-                              httpPost('unfollowrequest/${_profileData!['yalker_profile']['user_number']}', {"email":"email"},jwt: true);
+                              httpPost(
+                                  'unfollowrequest/${_profileData!['yalker_profile']['user_number']}',
+                                  {"email": "email"},
+                                  jwt: true);
                               _onPressedButton();
                             } catch (error) {
                               print('Error unfollow request: $error');
