+++ lib/search/search_post.dart
@@ -2,12 +2,12 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:yalkey_0206_test/yalker_profile_page.dart';
-import 'constant.dart';
+import '../profile/yalker_profile_page.dart';
+import '../constant.dart';
 import 'search_user.dart';
 import 'search_user_id.dart';
-import 'post/post_model.dart';
-import 'post/post_widget.dart';
+import '../post/post_model.dart';
+import '../post/post_widget.dart';
 
 class SearchPostListPage extends StatefulWidget {
   // 画面遷移元からのデータを受け取る変数
