+++ lib/setting/account_setting_list.dart
@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
-import 'login_page.dart';
+import '../login_page.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'profile_edit_page.dart';
-import 'deactivate_account.dart';
-import 'password_update.dart';
+import '../profile/profile_edit_page.dart';
+import '../setting/deactivate_account.dart';
+import '../setting/password_update.dart';
 import 'email_update.dart';
 import 'dart:convert';
-import 'api.dart';
-
+import '../api.dart';
 
 class AccountSettingListPage extends StatefulWidget {
   const AccountSettingListPage({Key? key}) : super(key: key);
@@ -16,10 +15,8 @@ class AccountSettingListPage extends StatefulWidget {
   _AccountSettingListPageState createState() => _AccountSettingListPageState();
 }
 
-
 class _AccountSettingListPageState extends State<AccountSettingListPage> {
   Map<String, dynamic>? _profileData;
-
 
   @override
   void initState() {
@@ -30,9 +27,8 @@ class _AccountSettingListPageState extends State<AccountSettingListPage> {
   Future<void> logout(context) async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
     await prefs.clear();
-    Navigator.pushAndRemoveUntil(
-        context,
-        MaterialPageRoute(builder: (context) => LoginPage()),(_) => false);
+    Navigator.pushAndRemoveUntil(context,
+        MaterialPageRoute(builder: (context) => LoginPage()), (_) => false);
   }
 
   Future<void> _fetchProfileData() async {
@@ -49,7 +45,7 @@ class _AccountSettingListPageState extends State<AccountSettingListPage> {
 
     try {
       final Map<String, dynamic> response =
-      await httpGet('login-user-profile/', jwt: true);
+          await httpGet('login-user-profile/', jwt: true);
       setState(() {
         _profileData = response;
       });
@@ -69,8 +65,6 @@ class _AccountSettingListPageState extends State<AccountSettingListPage> {
       ),
     );
   }
-
-
 
   @override
   Widget build(BuildContext context) {
@@ -138,4 +132,3 @@ class _AccountSettingListPageState extends State<AccountSettingListPage> {
     );
   }
 }
-
